Add files via upload
In this project I analyzed data about Airbnb accommodation in New York city
I did my project using Python jupyter notebook, the first step was a general overview of the data set then I had to check the null values, drop the unnecessary columns for the analyze purpose, in some cases the null values have to be fill because there are a lot in the same column in this case is a good task because not affect to the analyze, in this case was for the reviews per month column, next task was to remove the duplicated values.
Next step is to do different charts about the information from the data frame some examples of the analyze are about to count the airbnbs by neighbourhood_group, room_type and count the occurrences, distribution prices by room types, prices distribution by neighbourhood, prices distribution in the different neighbourhood Groups, average prices by neighbourhood_group according with the room type.
I also created some map visualizations of the different accommodation in the New York city area some examples of that are  map of the airbnbs by neighbourhood_group, map of the airbnbs by room types, map of the airbnbs by availability_365, map of prices less than the average price, map of prices more than the average price.
I analyzed the top of the busiest host using bar chart, analyze the correlation between the different variables, visualizations of the relationship between the columns, I created a barchart to view minimum nights by neighbourhood and the occurrences, check the availability by neighbourhood groups. 
Finally I did a lineal regression analysis where I found that the are not hard correlations of all the variables of these data, I check for the significance using the R-squared so I found that these variables are not very significant, according to the correlation analyzes the price and room type variables are the most correlated, but for the lineal regression I didn’t find an important lineal relation.